@@ -10,30 +10,27 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:http/http.dart' as http;
 
 class HomePageController {
-  static HomePageController _instance;
+  static HomePageController? _instance;
 
-  static getInstance() {
-    if (_instance == null) return HomePageController();
-    return _instance;
-  }
+  static getInstance() => _instance ?? HomePageController();
 
-  Future<DayStat> fetchTodaysStat() async {
+  Future<DayStat?> fetchTodaysStat() async {
     DateTime now = new DateTime.now();
     String dayId = now.year.toString() + dualNum(now.month) + dualNum(now.day);
     return await fetchDayStat(dayId);
   }
 
-  Future<DayStat> fetchYesterdaysStat() async {
+  Future<DayStat?> fetchYesterdaysStat() async {
     DateTime now = new DateTime.now();
     final yst = now.subtract(Duration(days: 1));
     String dayId = yst.year.toString() + dualNum(yst.month) + dualNum(yst.day);
     return await fetchDayStat(dayId);
   }
 
-  Future<DayStat> fetchDayStat(final String day) async {
-    FirebaseUser user = await Application.getInstance().getFirebaseUser();
+  Future<DayStat?> fetchDayStat(final String day) async {
+    User user = Application.getInstance().getFirebaseUser()!;
     final String url = AppString.apiDayStat + "${user.uid}/" + day;
-    final response = await http.get(url);
+    final response = await http.get(Uri.parse(url));
     if (response.statusCode == 200) {
       Map<String, dynamic> resMap = json.decode(response.body);
       if (resMap[AppString.keyStatus] == AppString.statusEmpty) {
@@ -45,26 +42,27 @@ class HomePageController {
     return null;
   }
 
-  Future<Definition> fetchDefinition(
+  Future<Definition?> fetchDefinition(
     final String word,
     final bool includeCall,
   ) async {
     final String query = includeCall
-        ? "?ic=True&uid=${(await Application.getInstance().getFirebaseUser()).uid}"
+        ? "?ic=True&uid=${(Application.getInstance().getFirebaseUser())!.uid}"
         : "";
     final String url = AppString.apiDefinition + '${word.toLowerCase()}/$query';
     print("Fetching def: $url");
-    var response = await http.get(url);
+    var response = await http.get(Uri.parse(url));
     if (response.statusCode == 200) return Definition.fromJson(response.body);
     return null;
   }
 
-  Future<DefinitionsResponse> fetchLearnedDefinitions({
-    final String url,
+  Future<DefinitionsResponse?> fetchLearnedDefinitions({
+    final String? url,
   }) async {
-    final uid = (await Application.getInstance().getFirebaseUser()).uid;
+    final uid = (Application.getInstance().getFirebaseUser())!.uid;
     final endpoint = url == null ? AppString.apiDefinitions + uid : url;
-    final res = await http.get(endpoint);
+    print(endpoint);
+    final res = await http.get(Uri.parse(endpoint));
     if (res.statusCode != 200) return null;
     final DefinitionsResponse defsRes = DefinitionsResponse.fromJson(res.body);
     return defsRes;
