@@ -1,24 +1,23 @@
 import 'dart:convert';
 
 class Sense {
-  String definition;
-  List<String> examples;
+  String? definition;
+  List<String>? examples;
 
   Sense({
     this.definition,
     this.examples,
   });
 
-  factory Sense.fromMap(Map<String, dynamic> map) {
+  static Sense? fromMap(Map<String, dynamic>? map) {
     if (map == null) return null;
-
     return Sense(
       definition: map['definition'],
       examples: List<String>.from(map['examples']),
     );
   }
 
-  factory Sense.fromJson(String source) => Sense.fromMap(json.decode(source));
+  static Sense? fromJson(String source) => Sense.fromMap(json.decode(source));
 
   Map<String, dynamic> toMap() {
     return {
