@@ -3,10 +3,13 @@ import 'package:bag_of_words/page/Homepage.dart';
 import 'package:bag_of_words/page/LoginPage.dart';
 import 'package:bag_of_words/res/AppColor.dart';
 import 'package:bag_of_words/res/AppString.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-void main() => runApp(MyApp());
+void main() {
+  runApp(MyApp());
+}
 
 class MyApp extends StatelessWidget {
   @override
@@ -28,14 +31,14 @@ class MyApp extends StatelessWidget {
 
   Widget _getInitialPage() {
     return FutureBuilder(
-      future: Application.getInstance().isAuthenticated(),
+      future: Firebase.initializeApp(),
       builder: (context, snapshot) {
         if (snapshot.connectionState != ConnectionState.done) {
           return Center(child: CircularProgressIndicator());
         }
 
-        final isAuthenticated = snapshot.data;
-        if (isAuthenticated != null && isAuthenticated) return HomePage();
+        final isAuthenticated = Application.getInstance().isAuthenticated();
+        if (isAuthenticated) return HomePage();
         return LoginPage();
       },
     );
