@@ -3,7 +3,7 @@ import 'package:bag_of_words/model/Definition.dart';
 class AppMockData {
   static String rowDefinition =
       '{"word": "mine","source_url": "https://www.oxfordlearnersdictionaries.com/definition/english/mine","origin": "Word Originpronoun Old English mīn, of Germanic origin; related to me (pronoun) and to Dutch mijn and German mein.","phonetics":[{"phonetic": "/maɪn/","audio": "https://www.oxfordlearnersdictionaries.com/media/english/uk_pron/m/min/mine_/mine__gb_1.mp3"},{"phonetic": "/maɪn/","audio": "https://www.oxfordlearnersdictionaries.com/media/english/us_pron/m/min/mine_/mine__us_1.mp3"}],"senses":[{"definition": "of or belonging to the person writing or speaking","examples": ["That\'s mine."," of mine He\'s a friend of mine (= one of my friends).","She wanted one like mine (= like I have)."]},{"definition": "my home","examples":["Let\'s go back to mine after the show."]}]}';
-  static Definition definition = Definition.fromJson(rowDefinition);
+  static Definition definition = Definition.fromJson(rowDefinition)!;
 
   static Map<String, dynamic> dayLearned = {
     "status": "SUCCESS",
